+++ client/models/NoteSchema.js
@@ -4,13 +4,10 @@ const Schema = mongoose.Schema;
 mongoose.promise = Promise;
 
 const NoteSchema = new Schema({
-
 	_id: Schema.Types.ObjectId,
-	// _id: new mongoose.Types.ObjectId(),
 	title: { type: String },
 	content: { type: String },
 	author: { type: Schema.Types.ObjectId, ref: 'User' }
-
 })
 
 NoteSchema.pre('save', function (next) {